requirements.txt
pyusb
pyee

if 'win' in sys.platform:
    pywin32
